This revision removes the OPT_LEVEL stuff (configure may report the optimization level, however, this no longer has any effect on the compiler flags.  I'm working on completely removing this from the build system).  I cleaned up some of the test code to accomodate use of -Werror.  Everything should compile cleanly now.  Several unused (I hope...) machine files have been removed, and I updated the Opteron flags to match the flash64 machine file.
main.cxx now uses fesetround to set the rounding mode (PPE builds only) based
on the compiler define PPE_ROUNDING_MODE.  Possible modes are:
FE_DOWNWARD, FE_TONEAREST, FE_TOWARDZERO, FE_UPWARD
The default is FE_TONEAREST if nothing is specified.

+++ accelerator/accelerator.cxx
@@ -23,8 +23,8 @@ int main(int argc, char *argv[]) {
 	// get an instance of the p2p connection
 	P2PConnection & p2p = P2PConnection::instance();
 
-	int rank = p2p.global_id();
-	int size = p2p.global_size();
+	size_t rank = p2p.global_id();
+	size_t size = p2p.global_size();
 
 	std::cout << "rank " << rank << " size " << size << std::endl;
